[FIX] Comment 관련 기능 추가 및 수정
유저 부분에 수정된 부분과 Comment관련 도메인 추가

+++ orury/src/main/java/org/kernel360/orury/board/comment/repository/CommentRepository.java
@@ -8,7 +8,5 @@ import java.util.List;
 
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Integer> {
-
-
-    List<Comment> findByPostId(Long postId);
+    List<Comment> findByPostId(Integer postId);
 }
